common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Local package to handle shared logic and schemas"
authors = [{ name="Caleb Northcott", email="crnorthc99@gmail.com" }]
dependencies = [
    "annotated-types==0.7.0",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "redis==6.2.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "python-dotenv==1.1.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"